refactor: combine sync and deploy scripts
