+++ scripts/deploy.ts
@@ -1,29 +1,77 @@
 import { select } from '@inquirer/prompts';
-import { spawn } from 'child_process';
-import { readFile, writeFile } from 'fs/promises';
+import matter from 'gray-matter';
+import { z } from 'zod';
+import { readFile, writeFile, readdir, stat, unlink, access, mkdir } from 'fs/promises';
+import { constants } from 'fs';
+import { spawn, SpawnOptions } from 'child_process';
+import { join, relative } from 'path';
+import { tmpdir } from 'os';
+import { randomUUID } from 'crypto';
 import { getRegistry } from '../src/lib/registry';
-import { ENV_KEYS, ENV_METADATA } from '../src/lib/env';
+import { env, ENV_KEYS, ENV_METADATA } from '../src/lib/env';
+import { INDEX_JSON, ROOT_JSON, INDEX_SLUG, SITEMAP_XML, SITEMAP_PAGES_XML } from '../src/lib/constants';
+import { PageMetadata, VaultDirectoryIndex, VaultRootIndex } from '../src/lib/vault';
 import { hasFlag, getFlagValue } from '../src/lib/cli';
+import { Registry, Site } from '../src/domain/registry';
 
-function execAsync(command: string, options: any = {}): Promise<void> {
+type CommandResult = {
+  status: number | null;
+  signal: NodeJS.Signals | null;
+  error?: Error;
+  stderr: Buffer;
+};
+
+type SpawnWithInputOptions = SpawnOptions & {
+  input?: string;
+};
+
+type RunMode = 'sync' | 'deploy' | 'configure';
+
+async function exists(path: string) {
+  try {
+    await access(path, constants.F_OK);
+    return true;
+  } catch {
+    return false;
+  }
+}
+
+async function execAsync({
+  command,
+  args,
+  options,
+}: {
+  command: string;
+  args: string[];
+  options: SpawnOptions;
+}): Promise<void> {
   return new Promise((resolve, reject) => {
-    const child = spawn(command, { ...options, shell: true });
+    // shell: false is more secure and robust against special characters
+    const child = spawn(command, args, { ...options, shell: false });
     child.on('close', (code) => {
       if (code === 0) resolve();
-      else reject(new Error(`Command failed with code ${code}`));
+      else reject(new Error(`Command "${command} ${args.join(' ')}" failed with code ${code}`));
     });
     child.on('error', (err) => reject(err));
   });
 }
 
-function spawnAsync(command: string, args: string[], options: any = {}): Promise<{ status: number | null, signal: string | null, error?: Error, stderr: Buffer }> {
+function spawnWithInput({
+  command,
+  args,
+  options = {},
+}: {
+  command: string;
+  args: string[];
+  options?: SpawnWithInputOptions;
+}): Promise<CommandResult> {
   return new Promise((resolve) => {
     const { input, ...spawnOptions } = options;
-    const child = spawn(command, args, spawnOptions);
+    const child = spawn(command, args, { ...spawnOptions, shell: false });
     let stderr = Buffer.alloc(0);
 
     if (child.stderr) {
-      child.stderr.on('data', (data) => {
+      child.stderr.on('data', (data: Buffer) => {
         stderr = Buffer.concat([stderr, data]);
       });
     }
@@ -43,97 +91,621 @@ function spawnAsync(command: string, args: string[], options: any = {}): Promise
   });
 }
 
-async function main() {
-  const isDev = hasFlag({ flag: '--dev' });
+const DateInputSchema = z.union([z.string(), z.number(), z.date()]);
 
-  // Check if vercel CLI is installed (only if NOT in dev mode)
-  if (!isDev) {
-    try {
-      await execAsync('vercel --version', { stdio: 'ignore' });
-    } catch {
-      console.error('⨯ Error: Vercel CLI is not installed or not in PATH.');
-      console.error('  Please install it with: npm install -g vercel');
-      process.exit(1);
+function parseSafeDate({
+  dateInput,
+  fallback,
+  label,
+  filePath,
+}: {
+  dateInput: unknown;
+  fallback: Date;
+  label: string;
+  filePath: string;
+}): string {
+  if (!dateInput) return fallback.toISOString();
+
+  const result = DateInputSchema.safeParse(dateInput);
+  if (!result.success) {
+    console.warn(
+      `⚠️  Warning: Invalid ${label} type for "${dateInput}" in ${filePath}. Falling back to file modification time.`
+    );
+    return fallback.toISOString();
+  }
+
+  const date = new Date(result.data);
+  if (isNaN(date.getTime())) {
+    console.warn(
+      `⚠️  Warning: Invalid ${label} value "${dateInput}" in ${filePath}. Falling back to file modification time.`
+    );
+    return fallback.toISOString();
+  }
+
+  return date.toISOString();
+}
+
+function escapeXml(unsafe: string): string {
+  const map: Record<string, string> = {
+    '&': '&amp;',
+    '<': '&lt;',
+    '>': '&gt;',
+    '"': '&quot;',
+    "'": '&apos;',
+  };
+  return unsafe.replace(/[&<>"']/g, (m) => map[m]);
+}
+
+function generateSitemapXml(urls: { loc: string; lastmod?: string }[]): string {
+  const urlTags = urls
+    .map(
+      (url) => `  <url>
+    <loc>${escapeXml(url.loc)}</loc>${url.lastmod ? `\n    <lastmod>${url.lastmod}</lastmod>` : ''}
+  </url>`
+    )
+    .join('\n');
+
+  return `<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+${urlTags}
+</urlset>`;
+}
+
+function generateSitemapIndexXml(sitemaps: { loc: string; lastmod?: string }[]): string {
+  const sitemapTags = sitemaps
+    .map(
+      (sitemap) => `  <sitemap>
+    <loc>${escapeXml(sitemap.loc)}</loc>${sitemap.lastmod ? `\n    <lastmod>${sitemap.lastmod}</lastmod>` : ''}
+  </sitemap>`
+    )
+    .join('\n');
+
+  return `<?xml version="1.0" encoding="UTF-8"?>
+<sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+${sitemapTags}
+</sitemapindex>`;
+}
+
+function mapPagesToSitemapUrls({
+  pages,
+  domain,
+  relDir = '',
+}: {
+  pages: PageMetadata[];
+  domain: string;
+  relDir?: string;
+}) {
+  return pages.map((p) => {
+    const urlPath = p.slug === INDEX_SLUG ? relDir : relDir ? `${relDir}/${p.slug}` : p.slug;
+    return {
+      loc: `https://${domain}/${urlPath}`,
+      lastmod: p.updatedAt || p.date,
+    };
+  });
+}
+
+async function writeSitemapFile({
+  fullPath,
+  relPath,
+  content,
+  dryRun,
+  label,
+}: {
+  fullPath: string;
+  relPath: string;
+  content: string;
+  dryRun: boolean;
+  label: string;
+}) {
+  if (!dryRun) {
+    const dir = join(fullPath, '..');
+    await mkdir(dir, { recursive: true });
+    await writeFile(fullPath, content);
+    console.log(`✨ Generated ${label} at ${relPath}`);
+  } else {
+    console.log(`[DRY RUN] Would generate ${label} at ${relPath}`);
+  }
+}
+
+async function scanPublicFiles({ dir, baseDir = dir }: { dir: string; baseDir?: string }): Promise<string[]> {
+  const files: string[] = [];
+
+  async function walk(currentDir: string) {
+    if (!(await exists(currentDir))) return;
+    const entries = await readdir(currentDir, { withFileTypes: true });
+
+    for (const entry of entries) {
+      const fullPath = join(currentDir, entry.name);
+
+      if (entry.isDirectory()) {
+        if (entry.name !== '.git' && entry.name !== 'node_modules') {
+          await walk(fullPath);
+        }
+      } else if (entry.isFile()) {
+        const relPath = relative(baseDir, fullPath);
+        files.push(relPath.replace(/\\/g, '/'));
+      }
     }
   }
 
-  // Parse registry path from command line arguments
-  const registryPath = getFlagValue({ flag: '--registry', alias: '-r' });
+  await walk(dir);
+  return files;
+}
 
-  const registry = await getRegistry(registryPath);
+async function generateAllSitemaps({
+  domain,
+  allIndices,
+  vaultPath,
+  dryRun,
+}: {
+  domain: string;
+  allIndices: Map<string, VaultDirectoryIndex>;
+  vaultPath: string;
+  dryRun: boolean;
+}): Promise<string[]> {
+  const subSitemaps: string[] = [];
+  for (const [relDir, indexData] of allIndices.entries()) {
+    if (relDir === '') continue; // Skip root, handled separately
+    if (indexData.pages.length === 0) continue;
 
-  const siteId = await select({
-    message: 'Select a site to deploy to Vercel:',
-    choices: registry.sites.map(site => ({
-      name: `${site.siteId} (${site.domain})`,
-      value: site.siteId,
-      description: `Project ID: ${site.vercelProjectId || 'Not configured'}`,
-    })),
+    const sitemapUrls = mapPagesToSitemapUrls({ pages: indexData.pages, domain, relDir });
+    const sitemapContent = generateSitemapXml(sitemapUrls);
+    const sitemapPath = join(vaultPath, 'public', relDir, SITEMAP_XML);
+    const relSitemapPath = `public/${relDir}/${SITEMAP_XML}`;
+
+    await writeSitemapFile({
+      fullPath: sitemapPath,
+      relPath: relSitemapPath,
+      content: sitemapContent,
+      dryRun,
+      label: `sitemap for "${relDir}"`,
+    });
+    subSitemaps.push(`${relDir}/${SITEMAP_XML}`);
+  }
+  return subSitemaps;
+}
+
+async function scanAndGenerate({
+  dir,
+  baseDir,
+  dryRun,
+  allIndices,
+}: {
+  dir: string;
+  baseDir: string;
+  dryRun: boolean;
+  allIndices: Map<string, VaultDirectoryIndex>;
+}): Promise<{ index: VaultDirectoryIndex; allDirs: string[] }> {
+  const entries = await readdir(dir, { withFileTypes: true });
+  const pages: PageMetadata[] = [];
+  let allDirs: string[] = [];
+
+  const relDir = relative(baseDir, dir).replace(/\\/g, '/');
+
+  for (const entry of entries) {
+    const fullPath = join(dir, entry.name);
+
+    if (entry.isDirectory()) {
+      if (entry.name !== '.git' && entry.name !== 'node_modules') {
+        const result = await scanAndGenerate({ dir: fullPath, baseDir, dryRun, allIndices });
+
+        const childRelDir = relative(baseDir, fullPath).replace(/\\/g, '/');
+        allDirs.push(childRelDir, ...result.allDirs);
+      }
+    } else if (entry.isFile() && entry.name.endsWith('.md')) {
+      const fileContent = await readFile(fullPath, 'utf-8');
+      const fileStats = await stat(fullPath);
+
+      const { data, content } = matter(fileContent);
+
+      if (data.published === false) {
+        continue;
+      }
+
+      const titleMatch = content.match(/^#\s+(.+)$/m);
+      const title =
+        typeof data.title === 'string' && data.title.trim() !== ''
+          ? data.title
+          : titleMatch
+          ? titleMatch[1].trim()
+          : entry.name;
+
+      const relPath = relative(dir, fullPath);
+      const slug = relPath.replace(/\.md$/, '').replace(/\\/g, '/');
+
+      const date = parseSafeDate({ dateInput: data.date, fallback: fileStats.mtime, label: 'date', filePath: relPath });
+      const manualUpdate = data.updated || data.lastmod;
+      const updatedAt = parseSafeDate({
+        dateInput: manualUpdate,
+        fallback: fileStats.mtime,
+        label: 'updated',
+        filePath: relPath,
+      });
+
+      const firstParagraph = content
+        .split('\n')
+        .map((line) => line.trim())
+        .filter((line) => line && !line.startsWith('#') && !line.startsWith('>'))[0];
+
+      let excerpt = '';
+      if (firstParagraph) {
+        excerpt = firstParagraph.slice(0, 160);
+        if (firstParagraph.length > 160) {
+          excerpt += '...';
+        }
+      }
+
+      pages.push({ title, slug, date, updatedAt, excerpt });
+    }
+  }
+
+  pages.sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime());
+
+  const indexData: VaultDirectoryIndex = {
+    version: 1,
+    pages,
+  };
+
+  const indexPath = join(dir, INDEX_JSON);
+  const relDirName = relDir || 'root';
+
+  if (!dryRun) {
+    await writeFile(indexPath, JSON.stringify(indexData, null, 2));
+    console.log(`✨ Generated index for "${relDirName}"`);
+  }
+
+  allIndices.set(relDir, indexData);
+
+  return { index: indexData, allDirs };
+}
+
+async function selectSite({
+  registry,
+  mode,
+  siteId,
+}: {
+  registry: Registry;
+  mode: RunMode;
+  siteId?: string;
+}): Promise<Site> {
+  const actionLabel = mode === 'sync' ? 'sync using AWS CLI' : mode === 'configure' ? 'configure locally' : 'sync and deploy';
+
+  const selectedSiteId =
+    siteId ||
+    (await select({
+      message: `Select a site to ${actionLabel}:`,
+      choices: registry.sites.map((site) => ({
+        name: `${site.siteId} (${site.domain || 'no domain'})`,
+        value: site.siteId,
+        description:
+          mode === 'configure'
+            ? `Project ID: ${site.vercelProjectId || 'Not configured'}`
+            : `Vault: ${site.vaultPath} -> Bucket: ${site.bucketName || 'Not configured'}`,
+      })),
+    }));
+
+  const site = registry.sites.find((s) => s.siteId === selectedSiteId);
+  if (!site) {
+    console.error(`⨯ Site "${selectedSiteId}" not found in registry.json`);
+    process.exit(1);
+  }
+
+  if (mode === 'configure') {
+    return site;
+  }
+
+  if (!site.bucketName) {
+    console.error(`⨯ Error: "bucketName" is not configured for site [${site.siteId}] in registry.json`);
+    process.exit(1);
+  }
+
+  if (!(await exists(site.vaultPath))) {
+    console.error(`⨯ Error: The local vaultPath does not exist: ${site.vaultPath}`);
+    process.exit(1);
+  }
+
+  return site;
+}
+
+function getEndpoint({ site, registry }: { site: Site; registry: Registry }): string | undefined {
+  return site.endpoint || registry.endpoint || env.S3_ENDPOINT;
+}
+
+function getS3Credentials({ registry }: { registry: Registry }) {
+  return {
+    accessKeyId: registry.accessKeyId || env.S3_ACCESS_KEY_ID,
+    secretAccessKey: registry.secretAccessKey || env.S3_SECRET_ACCESS_KEY,
+  };
+}
+
+async function syncSite({ site, registry, isDryRun }: { site: Site; registry: Registry; isDryRun: boolean }) {
+  const endpoint = getEndpoint({ site, registry });
+  const { accessKeyId, secretAccessKey } = getS3Credentials({ registry });
+
+  if (!endpoint || !accessKeyId || !secretAccessKey) {
+    console.error(
+      '⨯ Error: Missing S3 credentials. Please provide them in registry.json or via environment variables (S3_ENDPOINT, S3_ACCESS_KEY_ID, S3_SECRET_ACCESS_KEY).'
+    );
+    process.exit(1);
+  }
+
+  console.log(`\n☁️  Preparing AWS S3 Sync...`);
+  console.log(`- Local Path: ${site.vaultPath}`);
+  console.log(`- S3 Bucket:  ${site.bucketName}`);
+  console.log(`- Endpoint:   ${endpoint}\n`);
+
+  // We add a trailing slash to the vaultPath so that aws s3 sync syncs the *contents* of the directory
+  // and not the directory itself.
+  // Each site is synced to its own subdirectory in the bucket: /{site-id}/*
+  const args = [
+    's3',
+    'sync',
+    `${site.vaultPath}/`,
+    `s3://${site.bucketName}/${site.siteId}/`,
+    '--endpoint-url',
+    endpoint,
+    '--exclude',
+    '*.DS_Store',
+    '--exclude',
+    '*/.git/*',
+    '--exclude',
+    '.git/*',
+    '--delete',
+  ];
+
+  if (isDryRun) {
+    args.push('--dryrun');
+  }
+
+  console.log(`Executing:\n> aws ${args.join(' ')}\n`);
+
+  // stdio: 'inherit' passes the aws-cli output directly to our terminal
+  await execAsync({
+    command: 'aws',
+    args,
+    options: {
+      stdio: 'inherit',
+      env: {
+        ...process.env,
+        AWS_ACCESS_KEY_ID: accessKeyId,
+        AWS_SECRET_ACCESS_KEY: secretAccessKey,
+      },
+    },
+  });
+}
+
+async function uploadRegistry({ site, registry }: { site: Site; registry: Registry }) {
+  const endpoint = getEndpoint({ site, registry });
+  const { accessKeyId, secretAccessKey } = getS3Credentials({ registry });
+
+  if (!endpoint || !accessKeyId || !secretAccessKey) return;
+
+  console.log('\n✨ Uploading sanitized registry.json to bucket root...');
+
+  // Sanitize registry: remove sensitive credentials and local vault paths
+  const sanitizedSites = registry.sites
+    .filter((s) => s.bucketName === site.bucketName)
+    .map((s) => ({
+      domain: s.domain,
+      siteId: s.siteId,
+      // vaultPath is omitted or can be a placeholder
+    }));
+
+  const sanitizedRegistry = {
+    sites: sanitizedSites,
+  };
+
+  const registryTmpPath = join(tmpdir(), `notopress-registry-${randomUUID()}.json`);
+  try {
+    await writeFile(registryTmpPath, JSON.stringify(sanitizedRegistry, null, 2));
+
+    const args = ['s3', 'cp', registryTmpPath, `s3://${site.bucketName}/registry.json`, '--endpoint-url', endpoint];
+
+    await execAsync({
+      command: 'aws',
+      args,
+      options: {
+        stdio: 'inherit',
+        env: {
+          ...process.env,
+          AWS_ACCESS_KEY_ID: accessKeyId,
+          AWS_SECRET_ACCESS_KEY: secretAccessKey,
+        },
+      },
+    });
+  } finally {
+    if (await exists(registryTmpPath)) {
+      await unlink(registryTmpPath);
+    }
+  }
+}
+
+async function generateIndices({
+  vaultPath,
+  dryRun = false,
+}: {
+  vaultPath: string;
+  dryRun?: boolean;
+}): Promise<{ rootContentIndex: VaultDirectoryIndex; allIndices: Map<string, VaultDirectoryIndex> }> {
+  const contentDir = join(vaultPath, 'content');
+  if (!(await exists(contentDir))) {
+    console.error(`⨯ Error: The required "content" directory is missing in the vault: ${vaultPath}`);
+    process.exit(1);
+  }
+
+  // 1. Recursive generation of index.json for each level in content/
+  console.log(`\n🔍 Recursively scanning "content" directory in ${contentDir}...`);
+  const allIndices = new Map<string, VaultDirectoryIndex>();
+  const { index: rootContentIndex, allDirs } = await scanAndGenerate({
+    dir: contentDir,
+    baseDir: contentDir,
+    dryRun,
+    allIndices,
   });
 
-  const site = registry.sites.find(s => s.siteId === siteId);
-  if (!site) {
-    console.error('⨯ Site not found in registry.json');
-    process.exit(1);
+  const publicBaseDir = join(vaultPath, 'public');
+  if (!dryRun) {
+    await mkdir(publicBaseDir, { recursive: true });
   }
 
-  const vercelProjectId = site.vercelProjectId || site.siteId;
+  const publicFiles = (await exists(publicBaseDir)) ? await scanPublicFiles({ dir: publicBaseDir }) : [];
 
-  const endpoint = site.endpoint || registry.endpoint;
+  // 2. Generate root.json at vault root pointing to top-level content
+  const rootPath = join(vaultPath, ROOT_JSON);
+  const vaultRootIndex: VaultRootIndex = {
+    ...rootContentIndex,
+    directories: allDirs, // root.json contains the full directory map
+    publicFiles,
+  };
 
-  if (!endpoint) {
-    console.error(`⨯ Error: No S3 endpoint found. Please provide "endpoint" in site or registry configuration.`);
-    process.exit(1);
+  if (dryRun) {
+    console.log(`[DRY RUN] Would generate master root index at: ${rootPath}`);
+  } else {
+    await writeFile(rootPath, JSON.stringify(vaultRootIndex, null, 2));
+    console.log(`✨ Generated master root index with ${allDirs.length} directories at: ${rootPath}`);
   }
 
-  console.log(`\n🚀 Preparing deployment for ${site.domain}...`);
-  console.log(`- Site ID: ${site.siteId}`);
-  console.log(`- Vercel Project ID: ${vercelProjectId}${site.vercelProjectId ? '' : ' (fallback to siteId)'}`);
+  return { rootContentIndex, allIndices };
+}
 
-  const envVars = {
+async function generateSitemaps({
+  vaultPath,
+  domain,
+  rootContentIndex,
+  allIndices,
+  dryRun,
+}: {
+  vaultPath: string;
+  domain: string | undefined;
+  rootContentIndex: VaultDirectoryIndex;
+  allIndices: Map<string, VaultDirectoryIndex>;
+  dryRun: boolean;
+}) {
+  if (!domain) {
+    console.log('ℹ️  Skipping sitemap generation because "domain" is not configured in registry.json');
+    return;
+  }
+
+  const publicBaseDir = join(vaultPath, 'public');
+  const subSitemaps = await generateAllSitemaps({ domain, allIndices, vaultPath, dryRun });
+  const rootPages = rootContentIndex.pages;
+  const sitemapUrls = rootPages.length > 0 ? mapPagesToSitemapUrls({ pages: rootPages, domain }) : [];
+
+  if (subSitemaps.length === 0) {
+    // Case 1: No sub-sitemaps, write root pages to sitemap.xml directly
+    if (sitemapUrls.length > 0) {
+      const sitemapContent = generateSitemapXml(sitemapUrls);
+      const sitemapPath = join(publicBaseDir, SITEMAP_XML);
+      await writeSitemapFile({
+        fullPath: sitemapPath,
+        relPath: `public/${SITEMAP_XML}`,
+        content: sitemapContent,
+        dryRun,
+        label: 'sitemap',
+      });
+    }
+  } else {
+    // Case 2: Sub-sitemaps exist, sitemap.xml should be an index
+    const sitemaps = [];
+
+    if (sitemapUrls.length > 0) {
+      const sitemapContent = generateSitemapXml(sitemapUrls);
+      const sitemapPath = join(publicBaseDir, SITEMAP_PAGES_XML);
+      await writeSitemapFile({
+        fullPath: sitemapPath,
+        relPath: `public/${SITEMAP_PAGES_XML}`,
+        content: sitemapContent,
+        dryRun,
+        label: 'root pages sitemap',
+      });
+      sitemaps.push({ loc: `https://${domain}/${SITEMAP_PAGES_XML}` });
+    }
+
+    for (const subSitemap of subSitemaps) {
+      sitemaps.push({ loc: `https://${domain}/${subSitemap}` });
+    }
+
+    if (sitemaps.length > 0) {
+      const sitemapIndexContent = generateSitemapIndexXml(sitemaps);
+      const sitemapIndexPath = join(publicBaseDir, SITEMAP_XML);
+      await writeSitemapFile({
+        fullPath: sitemapIndexPath,
+        relPath: `public/${SITEMAP_XML}`,
+        content: sitemapIndexContent,
+        dryRun,
+        label: 'master sitemap index',
+      });
+    }
+  }
+}
+
+function getVercelProjectId({ site }: { site: Site }): string {
+  return site.vercelProjectId || site.siteId;
+}
+
+function getDeploymentEnvVars({ site, registry }: { site: Site; registry: Registry }): Record<string, string | undefined> {
+  return {
     [ENV_KEYS.S3_ACCESS_KEY_ID]: registry.accessKeyId,
     [ENV_KEYS.S3_SECRET_ACCESS_KEY]: registry.secretAccessKey,
-    [ENV_KEYS.S3_ENDPOINT]: endpoint,
+    [ENV_KEYS.S3_ENDPOINT]: getEndpoint({ site, registry }),
     [ENV_KEYS.S3_BUCKET]: site.bucketName,
     [ENV_KEYS.VAULT_ROOT]: site.siteId,
   };
+}
 
-  if (isDev) {
-    console.log(`\n🛠️  Running in DEV mode - Updating .env.local for ${site.siteId}...`);
+function getVercelCommandEnv({ projectId }: { projectId: string }): NodeJS.ProcessEnv {
+  return { ...process.env, VERCEL_PROJECT_ID: projectId };
+}
 
-    let envContent = '';
-    try {
-      envContent = await readFile('.env.local', 'utf-8');
-    } catch {
-      // .env.local might not exist, that's fine
-    }
-
-    const lines = envContent.split('\n');
-    const existingVars: Record<string, string> = {};
-
-    // Parse existing variables while preserving comments or structure is hard with a simple split,
-    // so we'll just parse keys and re-generate.
-    lines.forEach(line => {
-      const trimmed = line.trim();
-      if (trimmed && !trimmed.startsWith('#') && trimmed.includes('=')) {
-        const [key, ...valueParts] = trimmed.split('=');
-        existingVars[key.trim()] = valueParts.join('=').trim();
-      }
+async function ensureVercelCli() {
+  try {
+    await execAsync({
+      command: 'vercel',
+      args: ['--version'],
+      options: { stdio: 'ignore' },
     });
-
-    // Merge new values
-    const finalVars = { ...existingVars, ...envVars };
-
-    const newContent = Object.entries(finalVars)
-      .filter(([_, v]) => v !== undefined)
-      .map(([k, v]) => `${k}=${v}`)
-      .join('\n');
-
-    await writeFile('.env.local', newContent);
-    console.log('✅ .env.local updated successfully.');
-    console.log('\n✨ Local development context switched. Restart npm run dev to see changes.');
-    process.exit(0);
+  } catch {
+    console.error('⨯ Error: Vercel CLI is not installed or not in PATH.');
+    console.error('  Please install it with: npm install -g vercel');
+    process.exit(1);
   }
+}
+
+async function configureLocalEnvironment({ site, registry }: { site: Site; registry: Registry }) {
+  const envVars = getDeploymentEnvVars({ site, registry });
+
+  console.log(`\n🛠️  Updating .env.local for ${site.siteId}...`);
+
+  let envContent = '';
+  try {
+    envContent = await readFile('.env.local', 'utf-8');
+  } catch {
+    // .env.local is optional.
+  }
+
+  const existingVars: Record<string, string> = {};
+  for (const line of envContent.split('\n')) {
+    const trimmed = line.trim();
+    if (trimmed && !trimmed.startsWith('#') && trimmed.includes('=')) {
+      const [key, ...valueParts] = trimmed.split('=');
+      existingVars[key.trim()] = valueParts.join('=').trim();
+    }
+  }
+
+  const finalVars = { ...existingVars, ...envVars };
+  const newContent = Object.entries(finalVars)
+    .filter((entry): entry is [string, string] => entry[1] !== undefined)
+    .map(([key, value]) => `${key}=${value}`)
+    .join('\n');
+
+  await writeFile('.env.local', newContent);
+  console.log('✅ .env.local updated successfully.');
+  console.log('\n✨ Local development context switched. Restart npm run dev to see changes.');
+}
+
+async function syncVercelEnvironment({ site, registry }: { site: Site; registry: Registry }) {
+  const vercelProjectId = getVercelProjectId({ site });
+  const envVars = getDeploymentEnvVars({ site, registry });
 
   console.log(`\n📡 Synchronizing environment variables to Vercel...`);
 
@@ -143,63 +715,153 @@ async function main() {
       continue;
     }
 
-    const metadata = ENV_METADATA[key as keyof typeof ENV_KEYS];
+    const metadataKey = key as keyof typeof ENV_METADATA;
+    const metadata = ENV_METADATA[metadataKey];
     if (!metadata) {
       throw new Error(`Missing metadata for environment variable: ${key}`);
     }
-    const isSensitive = metadata.isSensitive;
-    const sensitiveFlag = isSensitive ? ['--sensitive'] : [];
 
-    // Diagnostic logging to help troubleshoot sensitivity issues
-    console.log(`  Syncing ${key}... (Sensitive: ${isSensitive})`);
+    const sensitiveFlag = metadata.isSensitive ? ['--sensitive'] : [];
+    console.log(`  Syncing ${key}... (Sensitive: ${metadata.isSensitive})`);
 
-    try {
-      // Step 1: Try to add the environment variable
-      const addResult = await spawnAsync('vercel', ['env', 'add', key, 'production', ...sensitiveFlag], {
+    const addResult = await spawnWithInput({
+      command: 'vercel',
+      args: ['env', 'add', key, 'production', ...sensitiveFlag],
+      options: {
         input: value,
         stdio: ['pipe', 'pipe', 'pipe'],
-        env: { ...process.env, VERCEL_PROJECT_ID: vercelProjectId }
-      });
+        env: getVercelCommandEnv({ projectId: vercelProjectId }),
+      },
+    });
 
-      if (addResult.status !== 0) {
-        const stderr = addResult.stderr.toString();
+    if (addResult.status !== 0) {
+      const stderr = addResult.stderr.toString();
 
-        // If it already exists, we use 'update' instead to avoid downtime
-        if (stderr.toLowerCase().includes('already exists')) {
-          const updateResult = await spawnAsync('vercel', ['env', 'update', key, 'production', ...sensitiveFlag], {
+      if (stderr.toLowerCase().includes('already exists')) {
+        const updateResult = await spawnWithInput({
+          command: 'vercel',
+          args: ['env', 'update', key, 'production', ...sensitiveFlag],
+          options: {
             input: value,
             stdio: ['pipe', 'inherit', 'inherit'],
-            env: { ...process.env, VERCEL_PROJECT_ID: vercelProjectId }
-          });
+            env: getVercelCommandEnv({ projectId: vercelProjectId }),
+          },
+        });
 
-          if (updateResult.error) throw updateResult.error;
-          if (updateResult.status !== 0) {
-            throw new Error(`Update failed with status ${updateResult.status}`);
-          }
-        } else {
-          // For any other error, we report it and fail
-          process.stderr.write(addResult.stderr);
-          throw addResult.error || new Error(`Command failed with status ${addResult.status}`);
+        if (updateResult.error) throw updateResult.error;
+        if (updateResult.status !== 0) {
+          throw new Error(`Update failed with status ${updateResult.status}`);
         }
+      } else {
+        process.stderr.write(addResult.stderr);
+        throw addResult.error || new Error(`Command failed with status ${addResult.status}`);
       }
-
-      console.log(`✅ ${key} synchronized${isSensitive ? ' (as sensitive)' : ''}.`);
-    } catch (err: any) {
-      console.error(`⨯ Failed to sync ${key}: ${err.message}`);
-      process.exit(1);
     }
+
+    console.log(`✅ ${key} synchronized${metadata.isSensitive ? ' (as sensitive)' : ''}.`);
   }
+}
+
+async function deployToVercel({ site }: { site: Site }) {
+  const vercelProjectId = getVercelProjectId({ site });
 
   console.log(`\n📦 Triggering production deployment...`);
-  try {
-    await execAsync('vercel deploy --prod', {
+  await execAsync({
+    command: 'vercel',
+    args: ['deploy', '--prod'],
+    options: {
       stdio: 'inherit',
-      env: { ...process.env, VERCEL_PROJECT_ID: vercelProjectId }
-    });
-    console.log(`\n✨ Deployment successfully triggered!`);
-  } catch (err: any) {
-    console.error(`\n⨯ Deployment failed.`);
-    console.error(err.message);
+      env: getVercelCommandEnv({ projectId: vercelProjectId }),
+    },
+  });
+  console.log(`\n✨ Deployment successfully triggered!`);
+}
+
+async function syncContent({
+  site,
+  registry,
+  isDryRun,
+}: {
+  site: Site;
+  registry: Registry;
+  isDryRun: boolean;
+}) {
+  const { rootContentIndex, allIndices } = await generateIndices({ vaultPath: site.vaultPath, dryRun: isDryRun });
+
+  await generateSitemaps({
+    vaultPath: site.vaultPath,
+    domain: site.domain,
+    rootContentIndex,
+    allIndices,
+    dryRun: isDryRun,
+  });
+
+  await syncSite({ site, registry, isDryRun });
+
+  if (!isDryRun) {
+    await uploadRegistry({ site, registry });
+  }
+}
+
+function getRunMode(): RunMode {
+  if (hasFlag({ flag: '--configure' }) || hasFlag({ flag: '--dev' })) {
+    return 'configure';
+  }
+
+  if (hasFlag({ flag: '--deploy' })) {
+    return 'deploy';
+  }
+
+  return 'sync';
+}
+
+async function main() {
+  const isDryRun = hasFlag({ flag: '--dry-run' });
+  const registryPath = getFlagValue({ flag: '--registry', alias: '-r' });
+  const siteId = getFlagValue({ flag: '--site', alias: '-s' });
+  const mode = getRunMode();
+
+  try {
+    const registry = await getRegistry(registryPath);
+
+    if (isDryRun) {
+      console.log('\n🏜️  DRY RUN MODE ENABLED - No changes will be made.');
+    }
+
+    const site = await selectSite({ registry, mode, siteId });
+
+    if (mode === 'configure') {
+      await configureLocalEnvironment({ site, registry });
+      return;
+    }
+
+    await syncContent({ site, registry, isDryRun });
+    if (isDryRun) {
+      console.log('\n✅ Dry run completed successfully!');
+      return;
+    }
+
+    console.log('\n✅ Sync and registry upload successfully completed!');
+
+    if (mode === 'deploy') {
+      const vercelProjectId = getVercelProjectId({ site });
+      console.log(`\n🚀 Preparing deployment for ${site.domain || site.siteId}...`);
+      console.log(`- Site ID: ${site.siteId}`);
+      console.log(`- Vercel Project ID: ${vercelProjectId}${site.vercelProjectId ? '' : ' (fallback to siteId)'}`);
+
+      await ensureVercelCli();
+      await syncVercelEnvironment({ site, registry });
+      await deployToVercel({ site });
+    }
+  } catch (err: unknown) {
+    const action = mode === 'deploy' ? 'Deployment' : mode === 'configure' ? 'Configuration' : isDryRun ? 'Dry run' : 'Sync process';
+    console.error(`\n⨯ ${action} failed.`);
+    if (err instanceof Error) {
+      console.error(err.message);
+    } else {
+      console.error('An unknown error occurred:');
+      console.error(err);
+    }
     process.exit(1);
   }
 }
